+++ src/api/index.js
@@ -6,11 +6,13 @@ import fetchChannels from './fetchChannels'
 import fetchOpeningHoursByType from './fetchOpeningHoursByType'
 import fetchOpeningHoursByRange from './fetchOpeningHoursByRange'
 import fetchStatus from './fetchStatus'
+import addDayActions from '../events/monthWidgetEvents'
 
 export default {
   fetchChannel,
   fetchChannels,
   ...fetchOpeningHoursByType,
   fetchOpeningHoursByRange,
-  fetchStatus
+  fetchStatus,
+  addDayActions
 }