+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2019-01-16
+### Added
+- polyfill for `Object.assign`
+- Click and keyboard events for the days in the month widget.
+
+### Updated
+- Move ajax methods to `./src/api` folder
+
 ## [0.0.3] - 2018-12-07
 ### Added
 - `pad` utility function which pads strings with a specified character